Broke out FriendlyJson Editor into smaller parts
